requirements.txt
opencv-python
numpy
requests
paho-mqtt
ultralytics
gTTS